done for today

+++ src/ECS/eventEnum.rs
@@ -1,7 +1,7 @@
 
 pub enum EventEnum{
     takeDamageEvent{id: usize, damage: usize},  // Id of enemy tower
-    towerAttackEvent{xTarget: usize, yTarget: usize, zTarget: usize},  
+    towerAttackEvent{xTarget: f32, yTarget: f32, zTarget: f32},  
     spawnTowerEvent{x: usize, y: usize, z: usize},
     spawnEnemyEvent,
     spawnProjectileEvent{x: usize, y: usize, z: usize},
